Wait for processes to be reaped if you are doing `rm -f`
If you do `rm -f` at just the right time when a container PID 1
has died but has not yet been reaped by Docker/Linux then kill
will return a ESRCH.  In this situation we wait a bit for
Docker/Linux to cleanup so we can delete the container.

As this is a race condition it is very difficult to reproduce but
can be done by writing a program that uses the Docker API and
does the equilavent of `docker rm -f $(docker run -d busybox echo hi)`.
Running that on the command line will never fail, but a program is
fast enough that it will hit situation a good amount of times.

+++ daemon/delete.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"os"
 	"path"
+	"syscall"
+	"time"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/docker/docker/engine"
@@ -55,7 +57,15 @@ func (daemon *Daemon) ContainerRm(job *engine.Job) engine.Status {
 		if container.IsRunning() {
 			if forceRemove {
 				if err := container.Kill(); err != nil {
-					return job.Errorf("Could not kill running container, cannot remove - %v", err)
+					if err != syscall.ESRCH {
+						return job.Errorf("Could not kill running container, cannot remove - %v", err)
+					}
+
+					// This is a situation in which the container PID 1 has died
+					// but has yet to be reaped or recognized by Docker as being gone.
+					if _, err := container.WaitStop(1 * time.Second); err != nil {
+						return job.Errorf("Failed to wait for killed container to stop, cannot remove - %v", err)
+					}
 				}
 			} else {
 				return job.Errorf("Conflict, You cannot remove a running container. Stop the container before attempting removal or use -f")
